Consolidate auth packages under internal/auth/
Auth concerns were scattered across three packages at different levels:

  internal/oauth/              -> internal/auth/oauth/
  internal/deviceflow/         -> internal/auth/device/
  internal/provider/googleauth/ -> internal/auth/google/

googleauth was never a provider — it's shared auth infrastructure
used by gemini and antigravity. deviceflow and oauth are also auth
infrastructure, not provider-specific. Grouping them under
internal/auth/ reflects the actual concern.
