+++ internal/provider/claude/response.go
@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 	"time"
 
+	"github.com/joshuadavidthomas/vibeusage/internal/auth/oauth"
 	"github.com/joshuadavidthomas/vibeusage/internal/models"
-	"github.com/joshuadavidthomas/vibeusage/internal/oauth"
 )
 
 // UsagePeriodResponse represents a single usage period from the Claude OAuth API.
