+++ internal/provider/copilot/copilot.go
@@ -7,12 +7,12 @@ import (
 	"os"
 	"time"
 
+	"github.com/joshuadavidthomas/vibeusage/internal/auth/device"
+	"github.com/joshuadavidthomas/vibeusage/internal/auth/oauth"
 	"github.com/joshuadavidthomas/vibeusage/internal/config"
-	"github.com/joshuadavidthomas/vibeusage/internal/deviceflow"
 	"github.com/joshuadavidthomas/vibeusage/internal/fetch"
 	"github.com/joshuadavidthomas/vibeusage/internal/httpclient"
 	"github.com/joshuadavidthomas/vibeusage/internal/models"
-	"github.com/joshuadavidthomas/vibeusage/internal/oauth"
 	"github.com/joshuadavidthomas/vibeusage/internal/provider"
 )
 
@@ -48,7 +48,7 @@ func (c Copilot) FetchStatus(ctx context.Context) models.ProviderStatus {
 // Auth returns the GitHub device flow for Copilot.
 func (c Copilot) Auth() provider.AuthFlow {
 	return provider.DeviceAuthFlow{
-		Config: deviceflow.Config{
+		Config: device.Config{
 			DeviceCodeURL: deviceCodeURL,
 			DeviceCodeParams: map[string]string{
 				"client_id": clientID,
